services/feature_pipeline/requirements.txt
flask
redis
prometheus_client
pyyaml
